# perf: avoid intermediate calls for Colors.FromInteger